@@ -1,11 +1,11 @@
 name: bitsdojo_window
 description: A package to help with creating custom windows with Flutter desktop (custom border, titlebar and minimize/maximize/close buttons) and common desktop window operations (show/hide/position on screen) for Windows and macOS
-version: 0.0.9
+version: 0.1.0
 homepage: https://www.bitsdojo.com
 repository: https://github.com/bitsdojo/bitsdojo_window
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 flutter:
@@ -21,11 +21,11 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  bitsdojo_window_platform_interface: ^0.0.2
+  bitsdojo_window_platform_interface: ^0.1.0
     #path: ../bitsdojo_window_platform_interface
-  bitsdojo_window_windows: ^0.0.3
+  bitsdojo_window_windows: ^0.1.0
     #path: ../bitsdojo_window_windows
-  bitsdojo_window_macos: ^0.0.3
+  bitsdojo_window_macos: ^0.1.0
     #path: ../bitsdojo_window_macos
-  bitsdojo_window_linux: ^0.0.1
+  bitsdojo_window_linux: ^0.1.0
     #path: ../bitsdojo_window_linux
