@@ -7,7 +7,7 @@ import 'package:flutter/widgets.dart';
 /// Close
 class CloseIcon extends StatelessWidget {
   final Color color;
-  CloseIcon({Key key, this.color}) : super(key: key);
+  CloseIcon({Key? key, required this.color}) : super(key: key);
   @override
   Widget build(BuildContext context) => Align(
         alignment: Alignment.topLeft,
@@ -28,7 +28,7 @@ class CloseIcon extends StatelessWidget {
 /// Maximize
 class MaximizeIcon extends StatelessWidget {
   final Color color;
-  MaximizeIcon({Key key, this.color}) : super(key: key);
+  MaximizeIcon({Key? key, required this.color}) : super(key: key);
   @override
   Widget build(BuildContext context) => _AlignedPaint(_MaximizePainter(color));
 }
@@ -46,8 +46,8 @@ class _MaximizePainter extends _IconPainter {
 class RestoreIcon extends StatelessWidget {
   final Color color;
   RestoreIcon({
-    Key key,
-    this.color,
+    Key? key,
+    required this.color,
   }) : super(key: key);
   @override
   Widget build(BuildContext context) => _AlignedPaint(_RestorePainter(color));
@@ -71,7 +71,7 @@ class _RestorePainter extends _IconPainter {
 /// Minimize
 class MinimizeIcon extends StatelessWidget {
   final Color color;
-  MinimizeIcon({Key key, this.color}) : super(key: key);
+  MinimizeIcon({Key? key, required this.color}) : super(key: key);
   @override
   Widget build(BuildContext context) => _AlignedPaint(_MinimizePainter(color));
 }
@@ -96,7 +96,7 @@ abstract class _IconPainter extends CustomPainter {
 }
 
 class _AlignedPaint extends StatelessWidget {
-  const _AlignedPaint(this.painter, {Key key}) : super(key: key);
+  const _AlignedPaint(this.painter, {Key? key}) : super(key: key);
   final CustomPainter painter;
 
   @override
