@@ -3,8 +3,8 @@ import '../app_window.dart';
 import 'package:flutter/foundation.dart' show kIsWeb;
 
 class _MoveWindow extends StatelessWidget {
-  _MoveWindow({Key key, this.child}) : super(key: key);
-  final Widget child;
+  _MoveWindow({Key? key, this.child}) : super(key: key);
+  final Widget? child;
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
@@ -18,22 +18,22 @@ class _MoveWindow extends StatelessWidget {
 }
 
 class MoveWindow extends StatelessWidget {
-  final Widget child;
-  MoveWindow({Key key, this.child}) : super(key: key);
+  final Widget? child;
+  MoveWindow({Key? key, this.child}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     if (child == null) return _MoveWindow();
     return _MoveWindow(
       child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
-          children: [Expanded(child: this.child)]),
+          children: [Expanded(child: this.child!)]),
     );
   }
 }
 
 class WindowTitleBarBox extends StatelessWidget {
-  final Widget child;
-  WindowTitleBarBox({Key key, this.child}) : super(key: key);
+  final Widget? child;
+  WindowTitleBarBox({Key? key, this.child}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     if (kIsWeb) {
