@@ -6,11 +6,14 @@ import 'package:ffi/ffi.dart';
 import 'package:win32/win32.dart';
 
 import './win32_plus.dart';
-import './native_api.dart';
+import './native_api.dart' as native;
 import 'package:bitsdojo_window_platform_interface/bitsdojo_window_platform_interface.dart';
 import './window_util.dart';
+import './window_interface.dart';
 
-bool isValidHandle(int handle, String operation) {
+var isInsideDoWhenWindowReady = false;
+
+bool isValidHandle(int? handle, String operation) {
   if (handle == null) {
     print("Could not $operation - handle is null");
     return false;
@@ -32,26 +35,35 @@ Rect getScreenRectForWindow(int handle) {
   return Rect.zero;
 }
 
-class WinWindow extends DesktopWindow {
-  int handle;
-  Size _minSize;
-  Size _maxSize;
-  Alignment _alignment;
+class WinWindow extends WinDesktopWindow {
+  int? handle;
+  Size? _minSize;
+  Size? _maxSize;
+  // We use this for reporting size inside doWhenWindowReady
+  // as GetWindowRect might not work reliably before window is shown on screen
+  Size? _sizeSetFromDart;
+  Alignment? _alignment;
+
+  void setWindowCutOnMaximize(int value) {
+    native.setWindowCutOnMaximize(value);
+  }
 
   WinWindow() {
     _alignment = Alignment.center;
   }
 
   Rect get rect {
+    if (!isValidHandle(handle, "get rectangle")) return Rect.zero;
     final winRect = calloc<RECT>();
-    GetWindowRect(handle, winRect);
+    GetWindowRect(handle!, winRect);
     Rect result = winRect.ref.toRect;
     calloc.free(winRect);
     return result;
   }
 
   set rect(Rect newRect) {
-    setWindowPos(handle, 0, newRect.left.toInt(), newRect.top.toInt(),
+    if (!isValidHandle(handle, "set rectangle")) return;
+    setWindowPos(handle!, 0, newRect.left.toInt(), newRect.top.toInt(),
         newRect.width.toInt(), newRect.height.toInt(), 0);
   }
 
@@ -62,6 +74,12 @@ class WinWindow extends DesktopWindow {
   }
 
   Size get sizeOnScreen {
+    if (isInsideDoWhenWindowReady == true) {
+      if (_sizeSetFromDart != null) {
+        final sizeOnScreen = getSizeOnScreen(_sizeSetFromDart!);
+        return sizeOnScreen;
+      }
+    }
     final winRect = this.rect;
     return Size(winRect.width, winRect.height);
   }
@@ -77,7 +95,8 @@ class WinWindow extends DesktopWindow {
   }
 
   int get dpi {
-    return GetDpiForWindow(handle);
+    if (!isValidHandle(handle, "get dpi")) return 96;
+    return GetDpiForWindow(handle!);
   }
 
   double get scaleFactor {
@@ -121,61 +140,77 @@ class WinWindow extends DesktopWindow {
     return Size(newWidth, newHeight);
   }
 
-  Alignment get alignment => _alignment;
+  Alignment? get alignment => _alignment;
 
   /// How the window should be aligned on screen
-  set alignment(Alignment newAlignment) {
+  set alignment(Alignment? newAlignment) {
     final sizeOnScreen = this.sizeOnScreen;
     _alignment = newAlignment;
-    final screenRect = getScreenRectForWindow(handle);
-    final rectOnScreen = getRectOnScreen(sizeOnScreen, _alignment, screenRect);
-    this.rect = rectOnScreen;
+    if (_alignment != null) {
+      if (!isValidHandle(handle, "set alignment")) return;
+      final screenRect = getScreenRectForWindow(handle!);
+      final rectOnScreen =
+          getRectOnScreen(sizeOnScreen, _alignment!, screenRect);
+      this.rect = rectOnScreen;
+    }
   }
 
-  set minSize(Size newSize) {
+  set minSize(Size? newSize) {
     _minSize = newSize;
-    setMinSize(_minSize.width.toInt(), _minSize.height.toInt());
+    if (newSize == null) {
+      //TODO - add handling for setting minSize to null
+      return;
+    }
+    native.setMinSize(_minSize!.width.toInt(), _minSize!.height.toInt());
   }
 
-  set maxSize(Size newSize) {
+  set maxSize(Size? newSize) {
     _maxSize = newSize;
-    setMaxSize(_maxSize.width.toInt(), _maxSize.height.toInt());
+    if (newSize == null) {
+      //TODO - add handling for setting maxSize to null
+      return;
+    }
+    native.setMaxSize(_maxSize!.width.toInt(), _maxSize!.height.toInt());
   }
 
   set size(Size newSize) {
+    if (!isValidHandle(handle, "set size")) return;
+
     var width = newSize.width;
 
-    if ((_minSize != null) && (newSize.width < _minSize.width)) {
-      width = _minSize.width;
+    if (_minSize != null) {
+      if (newSize.width < _minSize!.width) width = _minSize!.width;
     }
 
-    if ((_maxSize != null) && (newSize.width > _maxSize.width)) {
-      width = _maxSize.width;
+    if (_maxSize != null) {
+      if (newSize.width > _maxSize!.width) width = _maxSize!.width;
     }
 
     var height = newSize.height;
 
-    if ((_minSize != null) && (newSize.height < _minSize.height)) {
-      height = _minSize.height;
+    if (_minSize != null) {
+      if (newSize.height < _minSize!.height) height = _minSize!.height;
     }
 
-    if ((_maxSize != null) && (newSize.height > _maxSize.height)) {
-      height = _maxSize.height;
+    if (_maxSize != null) {
+      if (newSize.height > _maxSize!.height) height = _maxSize!.height;
     }
 
     Size sizeToSet = Size(width, height);
+    _sizeSetFromDart = sizeToSet;
     if (_alignment == null) {
-      SetWindowPos(handle, 0, 0, 0, sizeToSet.width.toInt(),
+      SetWindowPos(handle!, 0, 0, 0, sizeToSet.width.toInt(),
           sizeToSet.height.toInt(), SWP_NOMOVE);
     } else {
       final sizeOnScreen = getSizeOnScreen((sizeToSet));
-      final screenRect = getScreenRectForWindow(handle);
-      this.rect = getRectOnScreen(sizeOnScreen, _alignment, screenRect);
+      final screenRect = getScreenRectForWindow(handle!);
+      this.rect = getRectOnScreen(sizeOnScreen, _alignment!, screenRect);
     }
   }
 
   bool get isMaximized {
-    return (IsZoomed(handle) == 1);
+    if (!isValidHandle(handle, "get isMaximized")) return false;
+    return (IsZoomed(handle!) == 1);
   }
 
   @Deprecated("use isVisible instead")
@@ -184,7 +219,7 @@ class WinWindow extends DesktopWindow {
   }
 
   bool get isVisible {
-    return (IsWindowVisible(handle) == 1);
+    return (IsWindowVisible(handle!) == 1);
   }
 
   Offset get position {
@@ -193,21 +228,22 @@ class WinWindow extends DesktopWindow {
   }
 
   set position(Offset newPosition) {
-    SetWindowPos(handle, 0, newPosition.dx.toInt(), newPosition.dy.toInt(), 0,
+    if (!isValidHandle(handle, "set position")) return;
+    SetWindowPos(handle!, 0, newPosition.dx.toInt(), newPosition.dy.toInt(), 0,
         0, SWP_NOSIZE);
   }
 
   void show() {
     if (!isValidHandle(handle, "show")) return;
     setWindowPos(
-        handle, 0, 0, 0, 0, 0, SWP_NOSIZE | SWP_NOMOVE | SWP_SHOWWINDOW);
-    forceChildRefresh(handle);
+        handle!, 0, 0, 0, 0, 0, SWP_NOSIZE | SWP_NOMOVE | SWP_SHOWWINDOW);
+    forceChildRefresh(handle!);
   }
 
   void hide() {
     if (!isValidHandle(handle, "hide")) return;
     SetWindowPos(
-        handle, 0, 0, 0, 0, 0, SWP_NOSIZE | SWP_NOMOVE | SWP_HIDEWINDOW);
+        handle!, 0, 0, 0, 0, 0, SWP_NOSIZE | SWP_NOMOVE | SWP_HIDEWINDOW);
   }
 
   @Deprecated("use show()/hide() instead")
@@ -221,28 +257,28 @@ class WinWindow extends DesktopWindow {
 
   void close() {
     if (!isValidHandle(handle, "close")) return;
-    PostMessage(handle, WM_SYSCOMMAND, SC_CLOSE, 0);
+    PostMessage(handle!, WM_SYSCOMMAND, SC_CLOSE, 0);
   }
 
   void maximize() {
     if (!isValidHandle(handle, "maximize")) return;
-    PostMessage(handle, WM_SYSCOMMAND, SC_MAXIMIZE, 0);
+    PostMessage(handle!, WM_SYSCOMMAND, SC_MAXIMIZE, 0);
   }
 
   void minimize() {
     if (!isValidHandle(handle, "minimize")) return;
 
-    PostMessage(handle, WM_SYSCOMMAND, SC_MINIMIZE, 0);
+    PostMessage(handle!, WM_SYSCOMMAND, SC_MINIMIZE, 0);
   }
 
   void restore() {
     if (!isValidHandle(handle, "restore")) return;
-    PostMessage(handle, WM_SYSCOMMAND, SC_RESTORE, 0);
+    PostMessage(handle!, WM_SYSCOMMAND, SC_RESTORE, 0);
   }
 
   void maximizeOrRestore() {
     if (!isValidHandle(handle, "maximizeOrRestore")) return;
-    if (IsZoomed(handle) == 1) {
+    if (IsZoomed(handle!) == 1) {
       this.restore();
     } else {
       this.maximize();
@@ -251,7 +287,7 @@ class WinWindow extends DesktopWindow {
 
   set title(String newTitle) {
     if (!isValidHandle(handle, "set title")) return;
-    setWindowText(handle, newTitle);
+    setWindowText(handle!, newTitle);
   }
 
   void startDragging() {
