@@ -80,14 +80,14 @@ class RightSide extends StatelessWidget {
   }
 }
 
-const buttonColors = WindowButtonColors(
+final buttonColors = WindowButtonColors(
     iconNormal: Color(0xFF805306),
     mouseOver: Color(0xFFF6A00C),
     mouseDown: Color(0xFF805306),
     iconMouseOver: Color(0xFF805306),
     iconMouseDown: Color(0xFFFFD500));
 
-const closeButtonColors = WindowButtonColors(
+final closeButtonColors = WindowButtonColors(
     mouseOver: Color(0xFFD32F2F),
     mouseDown: Color(0xFFB71C1C),
     iconNormal: Color(0xFF805306),
