@@ -5,15 +5,13 @@ import 'package:bitsdojo_window_platform_interface/bitsdojo_window_platform_inte
 
 import './app_window.dart';
 import 'package:flutter/widgets.dart';
-import './native_api.dart';
 
 class BitsdojoWindowLinux extends BitsdojoWindowPlatform {
   BitsdojoWindowLinux() : super();
 
   @override
   void doWhenWindowReady(VoidCallback callback) {
-    WidgetsBinding.instance.waitUntilFirstFrameRasterized.then((value) {
-      setAppState(AppState.Ready);
+    WidgetsBinding.instance!.waitUntilFirstFrameRasterized.then((value) {
       callback();
     });
   }
