@@ -1,20 +1,24 @@
 library bitsdojo_window_windows;
 
 import 'dart:ui';
-import 'package:bitsdojo_window_platform_interface/bitsdojo_window_platform_interface.dart';
-
-import './app_window.dart';
 import 'package:flutter/widgets.dart';
+import 'package:bitsdojo_window_platform_interface/bitsdojo_window_platform_interface.dart';
+import './window.dart';
+import './app_window.dart';
 import './native_api.dart';
+
+export './window_interface.dart';
 
 class BitsdojoWindowWindows extends BitsdojoWindowPlatform {
   BitsdojoWindowWindows() : super();
 
   @override
   void doWhenWindowReady(VoidCallback callback) {
-    WidgetsBinding.instance.waitUntilFirstFrameRasterized.then((value) {
-      setAppState(AppState.Ready);
+    WidgetsBinding.instance!.waitUntilFirstFrameRasterized.then((value) {
+      isInsideDoWhenWindowReady = true;
+      setWindowCanBeShown(true);
       callback();
+      isInsideDoWhenWindowReady = false;
     });
   }
 
